requirements.txt
python-dotenv
spacy
mysql-connector-python
google-generativeai
